+++ tests/testthat/test_download_function.r
@@ -1,14 +1,19 @@
 # Phenocamr unit tests
 
+server_check <- phenocam_running(server_rois())
+
 # download routine
 test_that("check download options",{
 
+  # skip if server is down
+  skip_if_not(server_check)
+  
   # instigate once, needed on windows
   # checks for some reason otherwise
   # the connection is slow
   list_rois()
   
-  df = try(download_phenocam(site = "harvard$",
+  expect_message(download_phenocam(site = "bartlett$",
                          veg_type = "DB",
                          roi_id = "1000",
                          frequency = 1,
@@ -19,13 +24,13 @@ test_that("check download options",{
                          trim_daymet = FALSE,
                          out_dir = tempdir()))
 
-  df_raw = try(download_phenocam(site = "harvard$",
+  expect_message(download_phenocam(site = "bartlett$",
                              veg_type = "DB",
                              roi_id = "1000",
                              frequency = "roistats",
                              out_dir = tempdir()))
 
-  df_daymet = try(download_phenocam(site = "harvard$",
+  expect_message(download_phenocam(site = "bartlett$",
                              veg_type = "DB",
                              roi_id = "1000",
                              frequency = 3,
@@ -36,7 +41,7 @@ test_that("check download options",{
                              trim_daymet = TRUE,
                              out_dir = tempdir()))
 
-  df_contract = try(download_phenocam(site = "harvard$",
+  expect_message(download_phenocam(site = "bartlett$",
                                     veg_type = "DB",
                                     roi_id = "1000",
                                     frequency = 3,
@@ -47,7 +52,7 @@ test_that("check download options",{
                                     daymet = FALSE,
                                     out_dir = tempdir()))
 
-  df_false = try(download_phenocam(site = "harvard$",
+  df_false = try(download_phenocam(site = "bartlett$",
                              veg_type = "DB",
                              roi_id = "1000",
                              frequency = 3,
@@ -58,7 +63,7 @@ test_that("check download options",{
                              trim_daymet = FALSE,
                              out_dir = tempdir()))
   
-  df_internal = try(download_phenocam(site = "harvard$",
+  expect_message(download_phenocam(site = "bartlett$",
                                    veg_type = "DB",
                                    roi_id = "1000",
                                    frequency = 3,
@@ -70,14 +75,4 @@ test_that("check download options",{
                                    out_dir = tempdir(),
                                    internal = TRUE))
 
-  # see if any of the runs failed
-  check = !inherits(df,"try-error") &
-          !inherits(df_raw,"try-error") &
-          !inherits(df_false, "try-error") &
-          !inherits(df_contract, "try-error") &
-          !inherits(df_daymet,"try-error") &
-          !inherits(df_internal,"try-error")
-
-  # check if no error occured
-  expect_true(check)
 })